+++ src/vue/forms/SaleForm/components/ShortBlurbStepForm.vue
@@ -2,17 +2,20 @@
   <form
     novalidate
     class="app__form"
-    @submit.prevent="submit"
+    @submit.prevent="next"
   >
     <div class="app__form-row">
       <div class="app__form-field">
         <file-field
+          white-autofill
+          v-model="form.saleLogo"
+          @blur="touchField('form.saleLogo')"
+          @input="former.setAttr('saleLogo', form.saleLogo)"
+          name="create-sale-logo"
           :label="'create-sale-form.sale-logo-lbl' | globalize"
           :note="'create-sale-form.sale-logo-note' | globalize"
-          name="create-sale-logo"
           :file-extensions="['jpg', 'png']"
           :document-type="DOCUMENT_TYPES.saleLogo"
-          v-model="form.saleLogo"
           :error-message="getFieldErrorMessage('form.saleLogo')"
         />
       </div>
@@ -21,8 +24,10 @@
     <div class="app__form-row">
       <div class="app__form-field">
         <textarea-field
+          white-autofill
           v-model="form.shortDescription"
           @blur="touchField('form.shortDescription')"
+          @input="former.setAttr('shortDescription', form.shortDescription)"
           name="create-sale-short-description"
           :label="'create-sale-form.short-description-lbl' | globalize"
           :maxlength="DESCRIPTION_MAX_LENGTH"
@@ -50,38 +55,36 @@
 import FormMixin from '@/vue/mixins/form.mixin'
 
 import { DOCUMENT_TYPES } from '@/js/const/document-types.const'
-import { Document } from '@tokend/js-sdk'
 
-import { CreateSaleRequest } from '../wrappers/create-sale-request'
+import { required, maxLength, nonEmptyDocument } from '@validators'
+import { SaleFormer } from '@/js/formers/SaleFormer'
 
-import { required, maxLength } from '@validators'
-
-const EVENTS = {
-  submit: 'submit',
-}
 const DESCRIPTION_MAX_LENGTH = 255
 
 export default {
   name: 'short-blurb-step-form',
   mixins: [FormMixin],
   props: {
-    request: { type: CreateSaleRequest, default: null },
+    former: { type: SaleFormer, required: true },
   },
 
-  data: _ => ({
-    form: {
-      saleLogo: new Document(),
-      shortDescription: '',
-    },
-    DOCUMENT_TYPES,
-    DESCRIPTION_MAX_LENGTH,
-  }),
+  data () {
+    return {
+      form: {
+        saleLogo: this.former.attrs.saleLogo,
+        shortDescription: this.former.attrs.shortDescription,
+      },
+      DOCUMENT_TYPES,
+      DESCRIPTION_MAX_LENGTH,
+    }
+  },
 
   validations () {
     return {
       form: {
         saleLogo: {
           required,
+          nonEmptyDocument,
         },
         shortDescription: {
           required,
@@ -91,24 +94,9 @@ export default {
     }
   },
 
-  created () {
-    if (this.request) {
-      this.populateForm()
-    }
-  },
-
   methods: {
-    populateForm () {
-      this.form = {
-        saleLogo: new Document(this.request.logo),
-        shortDescription: this.request.shortDescription,
-      }
-    },
-
-    submit () {
-      if (this.isFormValid()) {
-        this.$emit(EVENTS.submit, this.form)
-      }
+    next () {
+      if (this.isFormValid()) this.$emit('next')
     },
   },
 }